+++ src/finlang/Transpiler/Transpiler.cs
@@ -109,7 +109,7 @@ public class Transpiler
         {
             INamedTypeSymbol symbol = model.GetDeclaredSymbol(classDeclNode).ThrowIfNull();
 
-            if (System.Diagnostics.Debugger.IsAttached && symbol.Name != "MemExample1")
+            if (System.Diagnostics.Debugger.IsAttached && symbol.Name != "QuadPoint")
                 continue;
 
             if (SymbolHelper.IsDerivedFrom(symbol, nameof(FinObj)) && !symbol.IsSimOnly())